WEB/requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
werkzeug.security
ItsDangerous==0.24
TextBlob
Tensorflow
Twint
